Add files via upload
My goal was to build an application that takes course information for a student, calculates the overall GPA, and displays the results.

To accomplish this, I declared and assigned values to variables of different data types, performed numeric operations, and used type casting to achieve accurate results. I also used the WriteLine() method and character escape sequences to format the output.